+++ include/stlsoft/containers/pod_vector.hpp
@@ -4,7 +4,7 @@
  * Purpose:     Contains the pod_vector class.
  *
  * Created:     23rd December 2003
- * Updated:     18th January 2024
+ * Updated:     28th January 2024
  *
  * Thanks to:   Chris Newcombe for requesting sufficient enhancements to
  *              auto_buffer such that pod_vector was born.
@@ -60,8 +60,8 @@
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_MAJOR       4
 # define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_MINOR       3
-# define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_REVISION    1
-# define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_EDIT        95
+# define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_REVISION    2
+# define STLSOFT_VER_STLSOFT_CONTAINERS_HPP_POD_VECTOR_EDIT        96
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 /* /////////////////////////////////////////////////////////////////////////
@@ -1264,7 +1264,7 @@ pod_vector<T_value, T_allocator, V_internalSize>::erase(
 {
     STLSOFT_ASSERT(is_valid_());
 
-    STLSOFT_ASSERT(first < end());
+    STLSOFT_ASSERT(first <= end());
     STLSOFT_ASSERT(!(first < begin()));
     STLSOFT_ASSERT(!(end() < last));
     STLSOFT_ASSERT(!(last < begin()));