+++ include/stlsoft/stlsoft.h
@@ -6,7 +6,7 @@
  *              types.
  *
  * Created:     15th January 2002
- * Updated:     16th January 2024
+ * Updated:     18th January 2024
  *
  * Home:        http://stlsoft.org/
  *
@@ -56,8 +56,8 @@
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define STLSOFT_VER_STLSOFT_H_STLSOFT_MAJOR    3
 # define STLSOFT_VER_STLSOFT_H_STLSOFT_MINOR    51
-# define STLSOFT_VER_STLSOFT_H_STLSOFT_REVISION 13
-# define STLSOFT_VER_STLSOFT_H_STLSOFT_EDIT     540
+# define STLSOFT_VER_STLSOFT_H_STLSOFT_REVISION 14
+# define STLSOFT_VER_STLSOFT_H_STLSOFT_EDIT     541
 #else /* ? STLSOFT_DOCUMENTATION_SKIP_SECTION */
 /* # include "./internal/doxygen_defs.h" */
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
@@ -341,12 +341,13 @@
 # define _STLSOFT_VER_1_10_3        0x010a03ff  /*!< Version 1.10.3 (12th January 2024) */
 # define _STLSOFT_VER_1_10_4        0x010a04ff  /*!< Version 1.10.4 (16th January 2024) */
 # define _STLSOFT_VER_1_10_5        0x010a05ff  /*!< Version 1.10.5 (18th January 2024) */
+# define _STLSOFT_VER_1_10_6        0x010a06ff  /*!< Version 1.10.6 (28th January 2024) */
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 #define _STLSOFT_VER_MAJOR      1
 #define _STLSOFT_VER_MINOR      10
-#define _STLSOFT_VER_REVISION   5
-#define _STLSOFT_VER            _STLSOFT_VER_1_10_5
+#define _STLSOFT_VER_REVISION   6
+#define _STLSOFT_VER            _STLSOFT_VER_1_10_6
 
 /* /////////////////////////////////////
  * underlying version detection