+++ prepare_cmake.sh
@@ -6,7 +6,9 @@ Basename=$(basename "$ScriptPath")
 CMakePath=$Dir/_build
 
 
-CmakeVerboseMakefile=0
+CMakeExamplesDisabled=0
+CMakeTestingDisabled=0
+CMakeVerboseMakefile=0
 Configuration=Release
 RunMake=0
 
@@ -16,23 +18,31 @@ RunMake=0
 
 while [[ $# -gt 0 ]]; do
     case $1 in
+        -v|--cmake-verbose-makefile)
+
+            CMakeVerboseMakefile=1
+            ;;
         -d|--debug-configuration)
 
             Configuration=Debug
+            ;;
+        -E|--disable-examples)
+
+            CMakeExamplesDisabled=1
+            ;;
+        -T|--disable-testing)
+
+            CMakeTestingDisabled=1
             ;;
         -m|--run-make)
 
             RunMake=1
             ;;
-        -v|--cmake-verbose-makefile)
-
-            CmakeVerboseMakefile=1
-            ;;
         --help)
 
             cat << EOF
 STLSoft is a suite of libraries that provide STL extensions and facades over operating-system and technology-specific APIs
-Copyright (c) 2019-2023, Matthew Wilson and Synesis Information Systems
+Copyright (c) 2019-2024, Matthew Wilson and Synesis Information Systems
 Copyright (c) 2002-2019, Matthew Wilson and Synesis Software
 Creates/reinitialises the CMake build script(s)
 
@@ -42,18 +52,29 @@ Flags/options:
 
     behaviour:
 
+    -v
+    --cmake-verbose-makefile
+        configures CMake to run verbosely (by setting
+        CMAKE_VERBOSE_MAKEFILE=ON)
+
     -d
     --debug-configuration
-        uses Debug configuration. Default is to use Release
+        use Debug configuration (by setting CMAKE_BUILD_TYPE=Debug). Default
+        is to use Release
+
+    -E
+    --disable-examples
+        disables building of examples (by setting BUILD_EXAMPLES=OFF)
+
+    -T
+    --disable-testing
+        disables building of tests (by setting BUILD_TESTING=OFF). This is
+        necessary, for example, when installing on a system that does not
+        (yet) have xTests - which itself depends on STLSOFT - installed
 
     -m
     --run-make
-        runs make after a successful running of Cmake
-
-    -v
-    --cmake-verbose-makefile
-        configures CMake to run verbosely (by setting CMAKE_VERBOSE_MAKEFILE
-        to be ON)
+        executes make after a successful running of CMake
 
 
     standard flags:
@@ -86,9 +107,18 @@ cd $CMakePath
 
 echo "Executing CMake"
 
-if [ $CmakeVerboseMakefile -eq 0 ]; then CmakeVerboseMakefileFlag="OFF" ; else CmakeVerboseMakefileFlag="ON" ; fi
+if [ $CMakeExamplesDisabled -eq 0 ]; then CMakeBuildExamplesFlag="ON" ; else CMakeBuildExamplesFlag="OFF" ; fi
 
-cmake -DCMAKE_VERBOSE_MAKEFILE:BOOL=$CmakeVerboseMakefileFlag -DCMAKE_BUILD_TYPE=$Configuration .. || (cd ->/dev/null ; exit 1)
+if [ $CMakeTestingDisabled -eq 0 ]; then CMakeBuildTestingFlag="ON" ; else CMakeBuildTestingFlag="OFF" ; fi
+
+if [ $CMakeVerboseMakefile -eq 0 ]; then CMakeVerboseMakefileFlag="OFF" ; else CMakeVerboseMakefileFlag="ON" ; fi
+
+cmake \
+    -DBUILD_EXAMPLES:BOOL=$CMakeBuildExamplesFlag \
+    -DBUILD_TESTING:BOOL=$CMakeBuildTestingFlag \
+    -DCMAKE_BUILD_TYPE=$Configuration \
+    -DCMAKE_VERBOSE_MAKEFILE:BOOL=$CMakeVerboseMakefileFlag \
+    .. || (cd ->/dev/null ; exit 1)
 
 status=0
 
@@ -103,7 +133,7 @@ fi
 
 cd ->/dev/null
 
-if [ $CmakeVerboseMakefile -ne 0 ]; then
+if [ $CMakeVerboseMakefile -ne 0 ]; then
 
     echo -e "contents of $CMakePath:"
     ls -al $CMakePath